Refactor: Move Toast logic to context and clean up errors
Extract Toast context and provider to a separate file. Remove unused useState from Toast component. Silently handle errors in BulletHell.
